+++ scripts/core/utilities/exports.js
@@ -1,3 +1,4 @@
+export { default as Cache }  from './cache.js';
 export { default as Colors } from './colors.js';
 export { default as Draw }   from './draw.js';
 export { default as Easing } from './easing.js';